config: move logger to new package
